Refactoring crumb-component and add title with icon
