+++ src/stories/containers/MakerDAOLegacyBudget/MakerDAOLegacyBudgetContainer.tsx
@@ -10,7 +10,8 @@ import LegacyIcon from './components/LegacyIcon';
 import { useMakerDAOLegacyBudget } from './useMakerDAOLegacyBudget';
 
 const MakerDAOLegacyBudget = () => {
-  const { handleChangeYearsEMakerDAOLegacyBudget, trailingAddress, year } = useMakerDAOLegacyBudget();
+  const { handleChangeYearsEMakerDAOLegacyBudget, trailingAddress, year, trailingAddressDesk, isMobile } =
+    useMakerDAOLegacyBudget();
   return (
     <PageContainer>
       <BreadcrumbYearNavigation
@@ -18,10 +19,15 @@ const MakerDAOLegacyBudget = () => {
         years={YEARS_FINANCES_SELECTED}
         handleChange={handleChangeYearsEMakerDAOLegacyBudget}
         selectedValue={year}
+        trailingAddressDesk={trailingAddressDesk}
+        title="MakerDAO Legacy Budget"
       />
       <Container>
         <ContainerTitle>
-          <IconTitle icon={<LegacyIcon />} title="MakerDAO Legacy Budget" />
+          <IconTitle
+            icon={<LegacyIcon width={isMobile ? 32 : 62} height={isMobile ? 32 : 62} />}
+            title="MakerDAO Legacy Budget"
+          />
         </ContainerTitle>
       </Container>
     </PageContainer>
